リファクタリング（破壊的変更）

+++ iij_coupon_checker.rb
@@ -50,70 +50,74 @@ Plugin.create(:iij_coupon_checker) do
           name: 'クーポンの使用を変更',
           condition: lambda { |_| true },
           visible: true,
-          role: :timeline) do |_|
+          role: :timeline) do |_opt|
     Plugin::IIJ_COUPON_CHECKER::CouponInfo.get_info.next { |data|
-      hdo_list = Hash.new
-      data.each do |d|
-        info = d.instance_variable_get(:@value)
-        code = info[:hdo_info].hdoServiceCode
-        status = info[:hdo_info].couponUse
-        hdo_list[code] = status
+      list = Hash.new
+      data.each do |info|
+        info[:hdo_info].each do |hdo|
+          key = hdo[:number] ? hdo[:number] : hdo[:hdoServiceCode]
+          list[key] = hdo[:hdoServiceCode]
+        end
       end
-      hdo_list
-    }.next { |list|
+
+      p list
+
       Delayer.new {
+        # クーポンのステータス
         status_list = %w(オン オフ)
+        # ダイアログ
         dialog = Gtk::Dialog.new('クーポンを切り替える',
                                  $main_application_window,
                                  Gtk::Dialog::DESTROY_WITH_PARENT,
                                  [Gtk::Stock::OK, Gtk::Dialog::RESPONSE_OK],
                                  [Gtk::Stock::CANCEL, Gtk::Dialog::RESPONSE_CANCEL])
-
+        # 切り替えるIDの選択肢の生成
         dialog.vbox.add(Gtk::Label.new('回線の種類'))
-        service_list = Gtk::ComboBox.new(true)
-        dialog.vbox.add(service_list)
-        list.each_key { |key| service_list.append_text(key) }
+        service = Gtk::ComboBox.new(true)
+        dialog.vbox.add(service)
+        list.each_key { |key| service.append_text(key) }
 
+        # オン/オフ切り替えボタン
         dialog.vbox.add(Gtk::Label.new('切り替え'))
         switch = Gtk::ComboBox.new(true)
         dialog.vbox.add(switch)
         status_list.each { |s| switch.append_text(s) }
 
-        service_list.set_active(0)
+        service.set_active(0)
         switch.set_active(0)
         dialog.show_all
-
         result = dialog.run
 
-        begin
-          Delayer::Deferred.fail(result) unless Gtk::Dialog::RESPONSE_OK
-          hdo = list.keys[service_list.active]
-          status = switch.active == 0 ? true : false
-
-          # SIMのクーポン状態と指定した状態（オン・オフ）が同じ場合はリクエストを行わない
-          if list.values[service_list.active] == status
-            msg = "サービスコード #{hdo} のクーポンのステータスはすでに#{status ? 'オン' : 'オフ'}です"
-            post(msg)
-          else
-            Thread.new {
-              Plugin::IIJ_COUPON_CHECKER::CouponInfo.switch(hdo, status).next { |_response|
-                msg = "クーポンのステータスが更新されました\n" +
-                    "hdoServiceCode: #{hdo}\n" +
-                    "クーポンのステータス: #{status ? 'オン' : 'オフ'}"
-                # 投稿
-                post(msg)
+        # ダイアログの結果に応じて処理を分ける
+        case result
+          when Gtk::Dialog::RESPONSE_OK
+            hdo = list.values[service.active]
+            status = !!(switch.active == 0)
+            # ダイアログを削除
+            dialog.destroy
+            Thread.new do
+              Plugin::IIJ_COUPON_CHECKER::CouponInfo.switch(hdo, status).next { |_|
+                notice 'coupon status successfully changed to %{status}' % {status: status ? 'オン' : 'オフ'}
+                post("クーポンのステータスが更新されました\nhdoServiceCode: %{hdo}\nクーポンのステータス: %{status}" % {hdo: hdo, status: status ? 'オン' : 'オフ'})
               }.trap { |e|
-                activity :iij_coupon_checker, "クーポンの切り替えに失敗しました: #{e}"
+                activity :iij_coupon_checker, 'クーポンの切り替えに失敗しました: %{error}' % {error: e.reason}
+                post("ステータスコード: %{status}\n%{reason}" % {status: e.status, reason: e.reason})
                 error e
-                post("ステータスコード: #{e.status} (#{e.reason})\n詳細: #{return_code}")
               }
-            }
-          end
-        ensure
-          dialog.destroy
+            end
+          when Gtk::Dialog::RESPONSE_CANCEL
+            # ダイアログを削除
+            dialog.destroy
+          else
+            # ダイアログを削除
+            dialog.destroy
+            Delayer::Deferred.fail(result)
         end
-      }.trap { |e| error e }
-    }.trap { |e| error e }
+      }
+    }.trap { |e|
+      activity :iij_coupon_checker, 'クーポン情報の取得に失敗しました: %{error}' % {error: e.reason}
+      error 'status: %{status}, reason: %{reason}' % {status: e.status, reason: e.reason}
+    }
   end
